deleted contactsSchemas
