+++ models/hooks.js
@@ -1,0 +1,4 @@
+export const handleSaveError = (error, data, next) => {
+  error.status = 400;
+  next();
+};
