+++ routes/contactsRouter.js
@@ -7,11 +7,11 @@ import {
   // updateContact,
   // deleteContact,
 } from "../controllers/contactsControllers.js";
-// import validateBody from "../helpers/validateBody.js";
-// import {
-//   createContactSchema,
-//   updateContactSchema,
-// } from "../schemas/contactsSchemas.js";
+import validateBody from "../helpers/validateBody.js";
+import {
+  createContactSchema,
+  // updateContactSchema,
+} from "../models/Contact.js";
 
 const contactsRouter = express.Router();
 
@@ -19,7 +19,7 @@ contactsRouter.get("/", getAllContacts);
 
 // contactsRouter.get("/:id", getOneContact);
 
-contactsRouter.post("/", createContact);
+contactsRouter.post("/", validateBody(createContactSchema), createContact);
 
 // contactsRouter.put("/:id", validateBody(updateContactSchema), updateContact);
